Fix: Extract DataProvider
