+++ test/Unit/Component/Image/ImageTest.php
@@ -88,7 +88,7 @@ class ImageTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider providerInvalidString
+     * @dataProvider Refinery29\Sitemap\Test\Unit\Component\DataProvider::providerInvalidString
      *
      * @param mixed $title
      */
@@ -101,29 +101,6 @@ class ImageTest extends \PHPUnit_Framework_TestCase
         $image = new Image($location);
 
         $image->withTitle($title);
-    }
-
-    /**
-     * @return \Generator
-     */
-    public function providerInvalidString()
-    {
-        $faker = $this->getFaker();
-
-        $values = [
-            null,
-            $faker->boolean(),
-            $faker->words,
-            $faker->randomNumber(),
-            $faker->randomFloat(),
-            new \stdClass(),
-        ];
-
-        foreach ($values as $value) {
-            yield [
-                $value,
-            ];
-        }
     }
 
     public function testWithTitleClonesObjectAndSetsValue()
@@ -142,7 +119,7 @@ class ImageTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider providerInvalidString
+     * @dataProvider Refinery29\Sitemap\Test\Unit\Component\DataProvider::providerInvalidString
      *
      * @param mixed $caption
      */
@@ -173,7 +150,7 @@ class ImageTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider providerInvalidString
+     * @dataProvider Refinery29\Sitemap\Test\Unit\Component\DataProvider::providerInvalidString
      *
      * @param mixed $geoLocation
      */
@@ -204,7 +181,7 @@ class ImageTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider providerInvalidString
+     * @dataProvider Refinery29\Sitemap\Test\Unit\Component\DataProvider::providerInvalidString
      *
      * @param mixed $licence
      */
